# swagger.json
{
...
  "info": {
    "description": "Comprehensive API for bus payment system with offers, rewards, bookings, and refunds",
    "title": "Bus Payments & Rewards API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/bookings/{booking_id}/modify": {
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "additional_services": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "modification_reason": {
                    "type": "string"
                  },
                  "new_booking_date": {
                    "format": "date-time",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Booking modified successfully"
          },
          "400": {
            "description": "Invalid modification request"
          },
          "404": {
            "description": "Booking not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Modify booking",
        "tags": [
          "Bookings"
        ]
      }
    },
    "/api/bookings/{booking_id}/status": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking status details"
          },
          "404": {
            "description": "Booking not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get booking status",
        "tags": [
          "Bookings"
        ]
      }
    },
...
    "/offers/hotel/book-hotel": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "check_in": {
                    "format": "date",
                    "type": "string"
                  },
                  "check_out": {
                    "format": "date",
                    "type": "string"
                  },
                  "customer_id": {
                    "type": "integer"
                  },
                  "guests": {
                    "items": {
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "hotel_id": {
                    "type": "string"
                  },
                  "rooms": {
                    "type": "integer"
                  }
                },
                "required": [
                  "hotel_id",
                  "customer_id",
                  "check_in",
                  "check_out",
                  "guests"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Hotel booked successfully"
          }
        },
        "summary": "Book hotel",
        "tags": [
          "Hotel Integration"
        ]
      }
    },
    "/offers/hotel/booking/{booking_reference}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "booking_reference",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel booking details"
          }
        },
        "summary": "Get hotel booking details",
        "tags": [
          "Hotel Integration"
        ]
      }
    },
...
    "/offers/hotel/search-hotels": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "check_in": {
                    "format": "date",
                    "type": "string"
                  },
                  "check_out": {
                    "format": "date",
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "guests": {
                    "default": 1,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "rooms": {
                    "default": 1,
                    "minimum": 1,
                    "type": "integer"
                  }
                },
                "required": [
                  "city",
                  "check_in",
                  "check_out"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Available hotels"
          }
        },
        "summary": "Search hotels",
        "tags": [
          "Hotel Integration"
        ]
      }
    },
    "/offers/search/travel-package": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "departure_date": {
                    "format": "date",
                    "type": "string"
                  },
                  "destination": {
                    "type": "string"
                  },
                  "origin": {
                    "type": "string"
                  },
                  "passengers": {
                    "minimum": 1,
                    "type": "integer"
                  },
                  "return_date": {
                    "format": "date",
                    "type": "string"
                  },
                  "rooms": {
                    "minimum": 1,
                    "type": "integer"
                  }
                },
                "required": [
                  "origin",
                  "destination",
                  "departure_date",
                  "return_date"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Available travel packages"
          }
        },
        "summary": "Search travel packages",
        "tags": [
          "Travel Integration"
        ]
      }
    },
...
    "/offers/travel/book-flight": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "customer_id": {
                    "type": "integer"
                  },
                  "flight_id": {
                    "type": "string"
                  },
                  "passengers": {
                    "items": {
                      "type": "object"
                    },
                    "type": "array"
                  }
                },
                "required": [
                  "flight_id",
                  "customer_id",
                  "passengers"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Flight booked successfully"
          }
        },
        "summary": "Book flight",
        "tags": [
          "Travel Integration"
        ]
      }
    },
    "/offers/travel/booking/{booking_reference}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "booking_reference",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking details"
          }
        },
        "summary": "Get travel booking details",
        "tags": [
          "Travel Integration"
        ]
      }
    },
    "/offers/travel/search-flights": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "departure_date": {
                    "format": "date",
                    "type": "string"
                  },
                  "destination": {
                    "type": "string"
                  },
                  "origin": {
                    "type": "string"
                  },
                  "passengers": {
                    "default": 1,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "return_date": {
                    "format": "date",
                    "type": "string"
                  }
                },
                "required": [
                  "origin",
                  "destination",
                  "departure_date"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Available flights"
          }
        },
        "summary": "Search flights",
        "tags": [
          "Travel Integration"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local development server",
      "url": "http://192.168.86.189:5001"
    }
  ]
}